pkl: fix overread by 1 byte in pk_cmd_get_next_match
2019-12-13  Dan Čermák  <dan.cermak@posteo.net>

	* src/pkl-env.c (hash_string): Avoid overflow of signed integers
	in hash calculation.

+++ src/pkl-env.c
@@ -60,7 +60,7 @@ hash_string (const char *name)
   len = strlen (name);
   hash = len;
   for (i = 0; i < len; i++)
-    hash = ((hash * 613) + (unsigned)(name[i]));
+    hash = ((hash * (size_t)613) + (unsigned)(name[i]));
 
 #define HASHBITS 30
   hash &= (1 << HASHBITS) - 1;
